Changed to support arbitrary precision with BigDecimal

+++ rest/src/main/java/com/wit/model/CalculationRequest.java
@@ -1,13 +1,15 @@
 package com.wit.model;
 
+import java.math.BigDecimal;
+
 public class CalculationRequest {
     private String operation;
-    private double a;
-    private double b;
+    private BigDecimal a;
+    private BigDecimal b;
 
     public CalculationRequest() {} // default constructor
 
-    public CalculationRequest(String operation, double a, double b) {
+    public CalculationRequest(String operation, BigDecimal a, BigDecimal b) {
         this.operation = operation;
         this.a = a;
         this.b = b;
@@ -17,9 +19,9 @@ public class CalculationRequest {
     public String getOperation() { return operation; }
     public void setOperation(String operation) { this.operation = operation; }
 
-    public double getA() { return a; }
-    public void setA(double a) { this.a = a; }
+    public BigDecimal getA() { return a; }
+    public void setA(BigDecimal a) { this.a = a; }
 
-    public double getB() { return b; }
-    public void setB(double b) { this.b = b; }
+    public BigDecimal getB() { return b; }
+    public void setB(BigDecimal b) { this.b = b; }
 }
